Ms-catalogo
Se termiron ciertas cosas del catalogo epro falta terminar el properities

+++ ms-catalogo/src/main/java/com/ms_catalogo/dto/PerfumeDto.java
@@ -1,5 +1,7 @@
 package com.ms_catalogo.dto;
 
+import java.math.BigDecimal;
+
 import com.ms_catalogo.model.Concentracion;
 import com.ms_catalogo.model.Genero;
 
@@ -8,9 +10,9 @@ public record PerfumeDto(
     String nombre,
     String marca,
     Integer ml,
-    Double precio,
+    BigDecimal precio,
     Concentracion concentracion,
     Genero genero,
-    String descripcion) {
-
-}
+    String descripcion)
+    
+    {}
